chat-service/requirements.txt
fastapi~=0.122.0
uvicorn~=0.35.0
openai~=1.93.0
pydantic~=2.12.5
mysql-connector-python~=9.5.0
chromadb~=1.3.5